consumer/requirements.txt
flask==3.1.3
pika==1.3.2
psycopg2-binary==2.9.10
redis==5.2.1